pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mech-server"
version = "0.1.0"
description = ""
authors = [ "Adamantios Zaras <adamantios.zaras@valory.xyz>",]
readme = "README.md"
repository = "https://github.com/valory-xyz/mech-tools-dev"
keywords = [ "mech", "tool", "developer", "autonomy", "autonomous", "autonolas", "olas", "service", "multi", "agent",]
classifiers = [ "Environment :: Console", "Environment :: Web Environment", "Development Status :: 2 - Pre-Alpha", "Intended Audience :: Developers", "License :: OSI Approved :: Apache Software License", "Natural Language :: English", "Operating System :: MacOS", "Operating System :: Microsoft", "Operating System :: Unix", "Programming Language :: Python :: 3.10", "Topic :: Communications", "Topic :: Internet", "Topic :: Scientific/Engineering", "Topic :: Software Development", "Topic :: System",]
[[tool.poetry.packages]]
include = "mtd"

[[tool.poetry.packages]]
include = "packages"

[[tool.poetry.include]]
path = "mtd/templates/runtime/*"
format = [ "sdist", "wheel",]

[[tool.poetry.include]]
path = "mtd/templates/runtime/.example.env"
format = [ "sdist", "wheel",]

[[tool.poetry.include]]
path = "mtd/templates/*.template"
format = [ "sdist", "wheel",]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
open-autonomy = "==0.21.11"
openai = "==1.30.2"
requests = "==2.28.2"
py-multibase = "==1.0.3"
py-multicodec = "==0.2.1"
grpcio = "==1.53.0"
asn1crypto = "<1.5.0,>=1.4.0"
open-aea-ledger-ethereum = "==2.0.8"
open-aea-ledger-cosmos = "*"
protobuf = "<4.25.0,>=4.21.6"
hypothesis = "==6.21.6"
open-aea-test-autonomy = "==0.21.11"
web3 = "<8,>=7.0.0"
ipfshttpclient = "==0.8.0a2"
open-aea-cli-ipfs = "==2.0.8"
pytest-asyncio = "*"
aiohttp = "<4.0.0,>=3.8.5"
certifi = "*"
multidict = "*"
ecdsa = ">=0.15"
eth_typing = "*"
hexbytes = "*"
packaging = "*"
py-ecc = "==8.0.0"
pytz = "==2022.2.1"
typing_extensions = "<=4.13.2,>=3.10.0.2"
websocket_client = "<1,>=0.32.0"
toml = "==0.10.2"
eth-abi = "==5.2.0"
pytest = "==7.4.4"
eth-utils = "==5.3.0"
pycryptodome = "==3.20.0"
openapi-core = "==0.15.0"
openapi-spec-validator = "<0.5.0,>=0.4.0"
click = "==8.1.8"
mech-client = "==0.18.8"
olas-operate-middleware = ">=0.14.16"
safe-eth-py = "^7.18.0"
mkdocs = "^1.6.1"
PyYAML = "==6.0.1"
werkzeug = "*"
attrs = "*"
prometheus_client = "==0.23.1"
pebble = "==5.1.3"

[tool.poetry.scripts]
mech = "mtd.cli:cli"

[tool.poetry.group.dev.dependencies]
setuptools = "^77.0.3"
mkdocs-material = "^9.6.17"
mkdocs-macros-plugin = "^1.3.9"
tomte = "==0.4.0"
